TIP-842: Move family normalizer to Akeneo\Pim\Structure
