+++ src/Akeneo/Pim/Structure/Component/Normalizer/Indexing/FamilyNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing;
+namespace Akeneo\Pim\Structure\Component\Normalizer\Indexing;
 
 use Akeneo\Channel\Component\Model\LocaleInterface;
 use Akeneo\Channel\Component\Repository\LocaleRepositoryInterface;
