+++ src/Akeneo/Pim/Structure/Component/Normalizer/Standard/FamilyNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Standard;
+namespace Akeneo\Pim\Structure\Component\Normalizer\Standard;
 
 use Akeneo\Pim\Structure\Component\Model\FamilyInterface;
 use Pim\Bundle\CatalogBundle\Filter\CollectionFilterInterface;
