+++ src/Pim/Bundle/CatalogBundle/Resources/config/serializers_indexing.yml
@@ -1,7 +1,6 @@
 parameters:
     pim_catalog.normalizer.indexing_product.product.class: Pim\Component\Catalog\Normalizer\Indexing\Product\ProductNormalizer
     pim_catalog.normalizer.indexing_product.product.properties.class: Pim\Component\Catalog\Normalizer\Indexing\Product\PropertiesNormalizer
-    pim_catalog.normalizer.indexing_product.product.family.class: Pim\Component\Catalog\Normalizer\Indexing\FamilyNormalizer
     pim_catalog.normalizer.indexing_product.datetime.class: Pim\Component\Catalog\Normalizer\Indexing\DateTimeNormalizer
     pim_catalog.normalizer.indexing_product.product.product_value_collection.class: Pim\Component\Catalog\Normalizer\Indexing\Value\ValueCollectionNormalizer
     pim_catalog.normalizer.indexing_product.product.completeness_collection.class: Pim\Component\Catalog\Normalizer\Indexing\CompletenessCollectionNormalizer
@@ -42,14 +41,6 @@ services:
         class: '%pim_catalog.normalizer.indexing_product.datetime.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.datetime'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_catalog.normalizer.indexing_product.product.family:
-        class: '%pim_catalog.normalizer.indexing_product.product.family.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.translation'
-            - '@pim_catalog.repository.locale'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
